refactor(gda-balancing): own attempted charge replay
